# Conflicts:
#	Assets/Graphics/Character/Meshes/Character_ver5.fbm.meta
#	Assets/Graphics/Mea.meta
#	Assets/Graphics/Mea/Animator.meta
#	Assets/Graphics/Mea/Prefabs.meta
#	Assets/Prefabs/Shimada.meta
#	ProjectSettings/Packages/com.unity.probuilder/Settings.json

+++ Assets/Scripts/Musashi/Core/StageData.cs
@@ -31,14 +31,14 @@ public class StageData
     [SerializeField,HideInInspector]
     private int m_highScore = 0;
     [SerializeField,HideInInspector]
-    private int m_clearTime = 0;
+    private float m_clearTime = 0;
     [SerializeField,HideInInspector]
     private bool m_isStageClear = false;
     #endregion
 
     #region SaveDataProperty
     public int HighScore { get => m_highScore; }
-    public int ClearTime { get => m_clearTime;}
+    public float ClearTime { get => m_clearTime;}
     public bool IsStageClear { get => m_isStageClear; }
     #endregion
 
@@ -50,7 +50,7 @@ public class StageData
         m_isStageClear = false;
     }
 
-    public void Save(int score, int claerTime)
+    public void Save(int score, float claerTime)
     {
         if (score > m_highScore)
         {
